ICU collation for strings #17.
Implemented using "golang.org/x/text/collate" package. A new
configuration API is exposed to set a collator object, which is
responsible for compiling strings to sort-key.

Note that SetTextCollator() API is not stable.

Another caveat, AFAIK, in using golang.org/x/text/collate is that
we cannot convert sort-key back to string.